frontend_ui/requirements.txt
fastapi
uvicorn
Pillow
requests
pydantic
stamina
tqdm
ollama
pdf2image
PyMuPDF
qdrant-client
python-multipart
pandas 
rouge-score 
nltk
scikit-learn
evaluate
anthropic
openai
streamlit
numpy
werkzeug
streamlit_lottie
aiohttp
psycopg2-binary